requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
buttons==0.1.9
cffi==1.14.6
chardet==4.0.0
git+https://github.com/Rapptz/discord-ext-menus@d9965ecbe2dba46c2b41cc9fd14aaa556841c183
discord.py==1.7.3
dnspython==1.16.0
idna==3.1
multidict==5.1.0
numpy==1.21.0
pandas==1.3.0
psutil==5.8.0
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2021.1
six==1.15.0
typing-extensions==3.7.4.3
tzlocal==2.1
yarl==1.6.3
